Fix shell provisioner error and echo behavior
There were two problems with the shell provisioner:

1. It kept running even if any particular step in the inline script failed
2. It didn't output the commands it ran even though this was desired by
   setting the `-x` flag in the shebang.

The reason for the first was a missing `-e` flag for the shell, and the reason
for the second was that while `-x` _was_ getting set, it was set only in the
shebang, but the shebang was being ignored because `execute_command` was
explicitly running a shell to execute the provisioner script.

This change fixes both problems by giving the `-x` and `-e` flags directly as
part of the `sh` command line args in `execute_command`.